deep_agent_research/requirements.txt
# 核心依赖
deepagents>=0.1.0
tavily-python>=0.1.0

# 开发依赖
pytest>=7.0.0
python-dotenv>=1.0.0